requirements.txt
tkinter
dnspython==2.3.0
pillow==10.0.0